Move TreeEnumerator to shared source
